+++ src/main/java/dev/tobee/telegram/request/body/SetChatAdministratorCustomTitleBody.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record SetChatAdministratorCustomTitleBody(
         @JsonProperty("chat_id") long chatId,
-        @JsonProperty("user_id") int userId,
+        @JsonProperty("user_id") long userId,
         @JsonProperty("custom_title") String customTitle
 ) {
 }
